adapt angles on clicks close to borders
when menu was open too close to a border, we were missing part of the menus.
This patch fixes this by adapting angle to the location of the client:
if we are too close to one of the parent widget border, we reduce angle.

For instance if we click next to top border, we only display menu items
on half of a circle, and if we are next to top/left border, we use a
quarter of a circle.